payload
this payload will allow you to have access to personal information on the user, on the computer and on the information of the installed program and disabilities the firewalls. It will save in addition to the information collected in a document inside in the computer the name of the document is: f: 145. this payload only works on windows systems and with a rubber ducky!